+++ affinda/operations/_affinda_api_operations.py
@@ -1072,6 +1072,56 @@ def build_delete_validation_result_request(
     )
 
 
+def build_batch_create_validation_results_request(
+    **kwargs  # type: Any
+):
+    # type: (...) -> HttpRequest
+    _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
+
+    content_type = kwargs.pop('content_type', _headers.pop('Content-Type', None))  # type: Optional[str]
+    accept = _headers.pop('Accept', "application/json")
+
+    # Construct URL
+    _url = kwargs.pop("template_url", "/v3/validation_results/batch_create")
+
+    # Construct headers
+    if content_type is not None:
+        _headers['Content-Type'] = _SERIALIZER.header("content_type", content_type, 'str')
+    _headers['Accept'] = _SERIALIZER.header("accept", accept, 'str')
+
+    return HttpRequest(
+        method="POST",
+        url=_url,
+        headers=_headers,
+        **kwargs
+    )
+
+
+def build_batch_delete_validation_results_request(
+    **kwargs  # type: Any
+):
+    # type: (...) -> HttpRequest
+    _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
+
+    content_type = kwargs.pop('content_type', _headers.pop('Content-Type', None))  # type: Optional[str]
+    accept = _headers.pop('Accept', "application/json")
+
+    # Construct URL
+    _url = kwargs.pop("template_url", "/v3/validation_results/batch_delete")
+
+    # Construct headers
+    if content_type is not None:
+        _headers['Content-Type'] = _SERIALIZER.header("content_type", content_type, 'str')
+    _headers['Accept'] = _SERIALIZER.header("accept", accept, 'str')
+
+    return HttpRequest(
+        method="POST",
+        url=_url,
+        headers=_headers,
+        **kwargs
+    )
+
+
 def build_get_all_extractors_request(
     **kwargs  # type: Any
 ):
@@ -1813,6 +1863,7 @@ def build_list_mapping_data_sources_request(
     limit = kwargs.pop('limit', _params.pop('limit', None))  # type: Optional[int]
     name = kwargs.pop('name', _params.pop('name', None))  # type: Optional[str]
     organization = kwargs.pop('organization', _params.pop('organization', None))  # type: Optional[str]
+    workspace = kwargs.pop('workspace', _params.pop('workspace', None))  # type: Optional[str]
     identifier = kwargs.pop('identifier', _params.pop('identifier', None))  # type: Optional[str]
     accept = _headers.pop('Accept', "application/json")
 
@@ -1828,6 +1879,8 @@ def build_list_mapping_data_sources_request(
         _params['name'] = _SERIALIZER.query("name", name, 'str')
     if organization is not None:
         _params['organization'] = _SERIALIZER.query("organization", organization, 'str')
+    if workspace is not None:
+        _params['workspace'] = _SERIALIZER.query("workspace", workspace, 'str')
     if identifier is not None:
         _params['identifier'] = _SERIALIZER.query("identifier", identifier, 'str')
 
@@ -3336,6 +3389,7 @@ def build_get_all_indexes_request(
     offset = kwargs.pop('offset', _params.pop('offset', None))  # type: Optional[int]
     limit = kwargs.pop('limit', _params.pop('limit', None))  # type: Optional[int]
     document_type = kwargs.pop('document_type', _params.pop('document_type', None))  # type: Optional[Union[str, "_models.Enum20"]]
+    name = kwargs.pop('name', _params.pop('name', None))  # type: Optional[str]
     accept = _headers.pop('Accept', "application/json")
 
     # Construct URL
@@ -3348,6 +3402,8 @@ def build_get_all_indexes_request(
         _params['limit'] = _SERIALIZER.query("limit", limit, 'int', maximum=100, minimum=1)
     if document_type is not None:
         _params['document_type'] = _SERIALIZER.query("document_type", document_type, 'str')
+    if name is not None:
+        _params['name'] = _SERIALIZER.query("name", name, 'str')
 
     # Construct headers
     _headers['Accept'] = _SERIALIZER.header("accept", accept, 'str')
@@ -6292,6 +6348,140 @@ class AffindaAPIOperationsMixin(object):  # pylint: disable=too-many-public-meth
 
     delete_validation_result.metadata = {"url": "/v3/validation_results/{id}"}  # type: ignore
 
+    def batch_create_validation_results(
+        self,
+        body,  # type: List[_models.ValidationResultCreate]
+        **kwargs,  # type: Any
+    ):
+        # type: (...) -> List[_models.ValidationResult]
+        """Batch create validation results.
+
+        Batch create validation results.
+
+        :param body:
+        :type body: list[~affinda.models.ValidationResultCreate]
+        :keyword callable cls: A custom type or function that will be passed the direct response
+        :return: list of ValidationResult, or the result of cls(response)
+        :rtype: list[~affinda.models.ValidationResult]
+        :raises: ~azure.core.exceptions.HttpResponseError
+        """
+        error_map = {
+            404: ResourceNotFoundError,
+            409: ResourceExistsError,
+            400: lambda response: HttpResponseError(
+                response=response, model=self._deserialize(_models.RequestError, response)
+            ),
+            401: lambda response: ClientAuthenticationError(
+                response=response, model=self._deserialize(_models.RequestError, response)
+            ),
+        }
+        error_map.update(kwargs.pop("error_map", {}) or {})
+
+        _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
+        _params = kwargs.pop("params", {}) or {}
+
+        content_type = kwargs.pop("content_type", _headers.pop("Content-Type", "application/json"))  # type: Optional[str]
+        cls = kwargs.pop("cls", None)  # type: ClsType[List[_models.ValidationResult]]
+
+        _json = self._serialize.body(body, "[ValidationResultCreate]")
+
+        request = build_batch_create_validation_results_request(
+            content_type=content_type,
+            json=_json,
+            template_url=self.batch_create_validation_results.metadata["url"],
+            headers=_headers,
+            params=_params,
+        )
+        request = _convert_request(request)
+        path_format_arguments = {
+            "region": self._serialize.url("self._config.region", self._config.region, "str"),
+        }
+        request.url = self._client.format_url(request.url, **path_format_arguments)  # type: ignore
+
+        pipeline_response = self._client._pipeline.run(  # type: ignore # pylint: disable=protected-access
+            request, stream=False, **kwargs
+        )
+        response = pipeline_response.http_response
+
+        if response.status_code not in [201]:
+            map_error(status_code=response.status_code, response=response, error_map=error_map)
+            error = self._deserialize.failsafe_deserialize(_models.RequestError, pipeline_response)
+            raise HttpResponseError(response=response, model=error)
+
+        deserialized = self._deserialize("[ValidationResult]", pipeline_response)
+
+        if cls:
+            return cls(pipeline_response, deserialized, {})
+
+        return deserialized
+
+    batch_create_validation_results.metadata = {"url": "/v3/validation_results/batch_create"}  # type: ignore
+
+    def batch_delete_validation_results(  # pylint: disable=inconsistent-return-statements
+        self,
+        body,  # type: _models.BatchDeleteValidationResultsRequest
+        **kwargs,  # type: Any
+    ):
+        # type: (...) -> None
+        """Batch delete validation results.
+
+        Batch delete validation results.
+
+        :param body:
+        :type body: ~affinda.models.BatchDeleteValidationResultsRequest
+        :keyword callable cls: A custom type or function that will be passed the direct response
+        :return: None, or the result of cls(response)
+        :rtype: None
+        :raises: ~azure.core.exceptions.HttpResponseError
+        """
+        error_map = {
+            404: ResourceNotFoundError,
+            409: ResourceExistsError,
+            400: lambda response: HttpResponseError(
+                response=response, model=self._deserialize(_models.RequestError, response)
+            ),
+            401: lambda response: ClientAuthenticationError(
+                response=response, model=self._deserialize(_models.RequestError, response)
+            ),
+        }
+        error_map.update(kwargs.pop("error_map", {}) or {})
+
+        _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
+        _params = kwargs.pop("params", {}) or {}
+
+        content_type = kwargs.pop("content_type", _headers.pop("Content-Type", "application/json"))  # type: Optional[str]
+        cls = kwargs.pop("cls", None)  # type: ClsType[None]
+
+        _json = self._serialize.body(body, "BatchDeleteValidationResultsRequest")
+
+        request = build_batch_delete_validation_results_request(
+            content_type=content_type,
+            json=_json,
+            template_url=self.batch_delete_validation_results.metadata["url"],
+            headers=_headers,
+            params=_params,
+        )
+        request = _convert_request(request)
+        path_format_arguments = {
+            "region": self._serialize.url("self._config.region", self._config.region, "str"),
+        }
+        request.url = self._client.format_url(request.url, **path_format_arguments)  # type: ignore
+
+        pipeline_response = self._client._pipeline.run(  # type: ignore # pylint: disable=protected-access
+            request, stream=False, **kwargs
+        )
+        response = pipeline_response.http_response
+
+        if response.status_code not in [204]:
+            map_error(status_code=response.status_code, response=response, error_map=error_map)
+            error = self._deserialize.failsafe_deserialize(_models.RequestError, pipeline_response)
+            raise HttpResponseError(response=response, model=error)
+
+        if cls:
+            return cls(pipeline_response, None, {})
+
+    batch_delete_validation_results.metadata = {"url": "/v3/validation_results/batch_delete"}  # type: ignore
+
     def get_all_extractors(
         self,
         organization,  # type: str
@@ -8066,6 +8256,7 @@ class AffindaAPIOperationsMixin(object):  # pylint: disable=too-many-public-meth
         limit=None,  # type: Optional[int]
         name=None,  # type: Optional[str]
         organization=None,  # type: Optional[str]
+        workspace=None,  # type: Optional[str]
         identifier=None,  # type: Optional[str]
         **kwargs,  # type: Any
     ):
@@ -8083,6 +8274,8 @@ class AffindaAPIOperationsMixin(object):  # pylint: disable=too-many-public-meth
         :type name: str
         :param organization: Filter by organization. Default value is None.
         :type organization: str
+        :param workspace: Filter by workspace. Default value is None.
+        :type workspace: str
         :param identifier: Filter by identifier. Default value is None.
         :type identifier: str
         :keyword callable cls: A custom type or function that will be passed the direct response
@@ -8114,6 +8307,7 @@ class AffindaAPIOperationsMixin(object):  # pylint: disable=too-many-public-meth
             limit=limit,
             name=name,
             organization=organization,
+            workspace=workspace,
             identifier=identifier,
             template_url=self.list_mapping_data_sources.metadata["url"],
             headers=_headers,
@@ -11716,6 +11910,7 @@ class AffindaAPIOperationsMixin(object):  # pylint: disable=too-many-public-meth
         offset=None,  # type: Optional[int]
         limit=None,  # type: Optional[int]
         document_type=None,  # type: Optional[Union[str, "_models.Enum20"]]
+        name=None,  # type: Optional[str]
         **kwargs,  # type: Any
     ):
         # type: (...) -> _models.PathsDvrcp3V3IndexGetResponses200ContentApplicationJsonSchema
@@ -11730,6 +11925,8 @@ class AffindaAPIOperationsMixin(object):  # pylint: disable=too-many-public-meth
         :type limit: int
         :param document_type: Filter indices by a document type. Default value is None.
         :type document_type: str or ~affinda.models.Enum20
+        :param name: Filter indices by name. Default value is None.
+        :type name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: PathsDvrcp3V3IndexGetResponses200ContentApplicationJsonSchema, or the result of
          cls(response)
@@ -11757,6 +11954,7 @@ class AffindaAPIOperationsMixin(object):  # pylint: disable=too-many-public-meth
             offset=offset,
             limit=limit,
             document_type=document_type,
+            name=name,
             template_url=self.get_all_indexes.metadata["url"],
             headers=_headers,
             params=_params,